# api-specification.yaml
openapi: 3.0.3
info:
  title: Turf Booking API
  description: REST APIs for turf booking mobile application
  version: 1.0.0
  contact:
    name: Hitendra Singh
    email: hitendras940@gmail.com
servers:
  - url: http://turfbackend-env.eba-yrja2qmi.ap-south-1.elasticbeanstalk.com
    description: Production server (AWS)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfileDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        phone:
          type: string
          example: "+919876543210"
        name:
          type: string
          example: "John Doe"
        memberSince:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
        totalBookings:
          type: integer
          example: 15
        currentBookings:
          type: integer
          example: 2
        completedBookings:
          type: integer
          example: 12
        cancelledBookings:
          type: integer
          example: 1
        totalLifetimeExpenditure:
          type: number
          format: double
          example: 45000.0
        pendingPayments:
          type: number
          format: double
          example: 3000.0
        currentBookingsList:
          type: array
          items:
            $ref: '#/components/schemas/BookingSummaryDto'
        previousBookingsList:
          type: array
          items:
            $ref: '#/components/schemas/BookingSummaryDto'
    BookingSummaryDto:
      type: object
      properties:
        bookingId:
          type: integer
          format: int64
          example: 1
        reference:
          type: string
          example: "TRF12AB34CD"
        turfName:
          type: string
          example: "Premium Cricket Ground"
        turfLocation:
          type: string
          example: "Sector 18, Noida"
        bookingDate:
          type: string
          format: date
          example: "2025-10-25"
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, REFUNDED]
          example: "CONFIRMED"
        amount:
          type: number
          format: double
          example: 4500.0
        slots:
          type: array
          items:
            type: object
            properties:
              startTime:
                type: string
                example: "09:00:00"
              endTime:
                type: string
                example: "10:00:00"
              price:
                type: number
                format: double
                example: 1500.0
        createdAt:
          type: string
          format: date-time
          example: "2025-10-20T10:30:00Z"
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: "Invalid or expired OTP"
        timestamp:
          type: string
          format: date-time
          example: "2025-10-18T12:00:00Z"
paths:
...
  /user/profile:
    get:
      tags:
        - User
      summary: Get user profile
      description: Retrieve the authenticated user's profile information
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDto'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden

    put:
      tags:
        - User
      summary: Update user profile
      description: Update the authenticated user's profile information
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "John Doe"
                phone:
                  type: string
                  example: "+919876543210"
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden